refactor: Header 추가 및 createHeader 함수 리팩토링
